mv testing helpers: distributing between tools and mocks.
