+++ express-zod-api/tests/unit/server.spec.ts
@@ -1,5 +1,5 @@
 import { moveRaw } from "../../src/server-helpers";
-import { givePort } from "../../../tests/helpers";
+import { givePort } from "../../../tools/ports";
 import {
   appMock,
   compressionMock,
